+++ android/port-test/src/com/zombieos/zsharp/porttest/GameActivity.java
@@ -1,0 +1,10 @@
+package com.zombieos.zsharp.porttest;
+
+import org.libsdl.app.SDLActivity;
+
+public final class GameActivity extends SDLActivity {
+    @Override
+    protected String[] getLibraries() {
+        return new String[] {"SDL3", "zsharp_android_vm"};
+    }
+}